Adding '-f' option for autoreconf.
Original source tarball may contain configure script which differs
in parts related to libtool and that leads to failure. Error message
says configure script should be rebuilt. Without -f autoreconf doesn't
rebuild configure from source tarball so -f is needed to do it.

The most probably that problem should be reported to libtool or
autoconf upstream.

This patch adds a workaround on that issue.